+++ Examples/Examples-iOS/IGListKitExamples/SectionControllers/GridSectionController.swift
@@ -20,11 +20,22 @@ final class GridItem: NSObject {
     let color: UIColor
     let itemCount: Int
 
+    var items: [String] = []
+
     init(color: UIColor, itemCount: Int) {
         self.color = color
         self.itemCount = itemCount
+
+        super.init()
+
+        self.items = computeItems()
     }
 
+    private func computeItems() -> [String] {
+        return [Int](1...itemCount).map {
+            String(describing: $0)
+        }
+    }
 }
 
 extension GridItem: ListDiffable {
@@ -42,8 +53,10 @@ extension GridItem: ListDiffable {
 final class GridSectionController: ListSectionController {
 
     private var object: GridItem?
+    private let isReorderable: Bool
 
-    override init() {
+    required init(isReorderable: Bool = false) {
+        self.isReorderable = isReorderable
         super.init()
         self.minimumInteritemSpacing = 1
         self.minimumLineSpacing = 1
@@ -63,7 +76,7 @@ final class GridSectionController: ListSectionController {
         guard let cell = collectionContext?.dequeueReusableCell(of: CenterLabelCell.self, for: self, at: index) as? CenterLabelCell else {
             fatalError()
         }
-        cell.text = "\(index + 1)"
+        cell.text = object?.items[index] ?? "undefined"
         cell.backgroundColor = object?.color
         return cell
     }
@@ -72,4 +85,13 @@ final class GridSectionController: ListSectionController {
         self.object = object as? GridItem
     }
 
+    override func canMoveItem(at index: Int) -> Bool {
+        return isReorderable
+    }
+
+    override func moveObject(from sourceIndex: Int, to destinationIndex: Int) {
+        guard let object = object else { return }
+        let item = object.items.remove(at: sourceIndex)
+        object.items.insert(item, at: destinationIndex)
+    }
 }