+++ Examples/Examples-iOS/IGListKitExamples/ViewControllers/DemosViewController.swift
@@ -58,7 +58,11 @@ final class DemosViewController: UIViewController, ListAdapterDataSource {
         DemoItem(name: "Calendar (auto diffing)",
                  controllerClass: CalendarViewController.self),
         DemoItem(name: "Dependency Injection",
-                 controllerClass: AnnouncingDepsViewController.self)
+                 controllerClass: AnnouncingDepsViewController.self),
+        DemoItem(name: "Reorder Cells",
+                 controllerClass: ReorderableViewController.self),
+        DemoItem(name: "Reorder Stacked Section Controllers",
+                 controllerClass: ReorderableStackedViewController.self)
     ]
 
     override func viewDidLoad() {